Basic .tac file serving static content over HTTP from the ./static directory
The .tac file is the heart of a twisted application it describes how twistd
is supposed to run our application.

The common pattern that can be observed here is the server, protocol and resource
combination used to create an application. Where TCPServer is the server,
Site is the HTTP protocol factory and File is the resource.

Separation of this layers opens up the possibility of switching out all parts
of the service with alternative implementations.

To run the static webserver described in the tac file run the following command:
 $ twistd -ny shutter.tac

The flags we use tell twisted not to daemonize (-n) the application and to read
the application (-y) variable from our .tac file.

The variable name `application` is a mandatory part of a .tac file.