We now introduce a stub service just for the snapshot list
RemoteShutterFactory will initalise instances of our ShutterJSONRPCProtocol.

For now the only implemented method is jsonrpc_snapshots which expects a json call in the format of:
{ "id" : 42, "method" : "snapshots", "params" : ["http://programistamag.pl"] }

and will provide a response containing url addresses of all snapshots for the provided url:
{ "id" : 42, "error" : null, "result" : [] }

Currently the method always returns an empty list but this stub will allow us to write
unit tests for the application and layer on more functionality as we go on.

We also introduce a second change to the .tac file. Using MultiService allows us to combine
more then one service in our application. When started our app will now run both a static
HTTP file server on port 8000 and our jsonrpc protocol on port 8001.